python/pyproject.toml
[project]
name = "resolveagent"
version = "0.1.0"
description = "ResolveAgent Runtime - Mega Agent Platform"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "AI Guru Global", email = "dev@resolveagent.io"},
]
keywords = ["agent", "llm", "ai", "fta", "rag", "workflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
    "agentscope>=0.1.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "protobuf>=5.28.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
resolveagent-docsync = "resolveagent.docsync.cli:main"
kudig-rag-import = "resolveagent.corpus.kudig_rag_import:main"
vectorize-rag-seeds = "resolveagent.corpus.seed_vectorizer:main"

[project.optional-dependencies]
rag = [
    "pymilvus>=2.4.0",
    "qdrant-client>=1.12.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resolveagent"]

[tool.ruff]
target-version = "py311"
line-length = 150

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["UP006", "UP035"]

[tool.ruff.lint.per-file-ignores]
"src/resolveagent/corpus/seed_vectorizer.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = false
ignore_missing_imports = true
warn_unused_ignores = false
disable_error_code = ["attr-defined", "call-arg", "assignment", "return-value", "arg-type", "union-attr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
